Move workspace setup status into the workspace banner
The workspace setup command's run status (running/succeeded/failed, logs,
cancel/rerun) was rendered in the per-agent chat intro. That had two problems:

- Terminal agents have no chat intro, so a setup run — including a failure —
  was completely invisible to them.
- In chat, the card permanently occupied space at the top of the agent pane
  instead of scrolling away like other intro content.

Setup is a workspace concern, not a per-agent one: there is a single run per
workspace, gated on workspace state rather than agent type, and every agent in
the workspace shares it. Surface it once in the WorkspaceBanner (workspace-level
chrome shown above both the chat and terminal panels) instead.

- Add WorkspaceSetupStatus: a compact banner pill (status badge + inline
  cancel/rerun/edit) with a logs popover. Renders only once a run exists
  (pending/running/succeeded/failed/legacy); participates in the banner's
  progressive-collapse at priority 3.
- Slim SetupStatusCard to the chat-intro onboarding affordance only (the
  configure-a-command CTA, and the one-click Run-setup row for workspaces
  configured after creation). Run states are now owned by the banner.
- Extract useSetupCommandActions for the shared cancel/rerun/edit handlers and
  command resolution so the two surfaces can't drift.

All setup-* test ids are preserved, and the chat-intro card and banner pill are
mutually exclusive by state, so the existing page-wide test helpers keep
working. Adds test_terminal_agent_surfaces_setup_status.
