+++ sculptor/frontend/src/pages/workspace/components/WorkspaceSetupStatus.module.scss
@@ -1,0 +1,66 @@
+// Compact setup-status segment for the WorkspaceBanner. Adopts the banner's
+// branch-segment idiom (small inline-flex chip, gray-a3 hover) so it reads as
+// one more banner item alongside the branch and diff summary.
+
+.pill {
+  align-items: center;
+  border-radius: var(--radius-1);
+  color: var(--gray-11);
+  cursor: pointer;
+  display: inline-flex;
+  flex-shrink: 0;
+  gap: var(--space-1);
+  padding: 2px 6px;
+  transition: background var(--duration-fast) ease;
+
+  &:hover,
+  &:focus-visible {
+    background: var(--gray-a3);
+    outline: none;
+  }
+}
+
+.pillOpen,
+.pillOpen:hover {
+  background: var(--accent-a3);
+}
+
+.pillInert {
+  cursor: default;
+
+  &:hover,
+  &:focus-visible {
+    background: transparent;
+  }
+}
+
+.pillError {
+  color: var(--red-11);
+}
+
+.icon {
+  color: var(--gray-11);
+  flex-shrink: 0;
+}
+
+.pillError .icon {
+  color: var(--red-11);
+}
+
+.label {
+  color: inherit;
+  font-size: var(--font-size-1);
+  white-space: nowrap;
+}
+
+.truncationBanner {
+  background: var(--amber-a3);
+  color: var(--amber-11);
+  font-size: var(--font-size-1);
+  padding: var(--space-1) var(--space-3);
+}
+
+.popoverEmpty {
+  color: var(--gray-9);
+  font-style: italic;
+}
